# backend/doc/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Social_App API (from PettersonApps)",
    "description": "Create REST API. Implement authorization and CRUD for users ",
    "contact": {
      "name": "Dmytro",
      "url": "https://github.com/dimanno",
      "email": "dimanno.dev@gmail.com"
    },
    "version": "1.1"
  },
  "servers": [
    {
      "url": "http://localhost:5050"
    }
  ],
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Returns a list of users",
        "description": "getUsers",
        "responses": {
          "200": {
            "description": "All users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserBodyModels"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Just an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/commonErrors"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create user",
        "description": "user object that needs to be added",
        "operationId": "addUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "name",
                  "username",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "default": "admin"
                  },
                  "username": {
                    "type": "string",
                    "default": "adminTest"
                  },
                  "email": {
                    "type": "string",
                    "default": "admin@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "default": "12345Az$"
                  },
                  "city": {
                    "type": "string",
                    "default": "Lviv"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBodyModels"
                }
              }
            }
          },
          "400": {
            "description": "Just an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/commonErrors"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserBodyModels": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "MongoDB ID",
            "example": "12345qwerty"
          },
          "email": {
            "type": "string",
            "example": "admin@gmail.com"
          },
          "name": {
            "type": "string",
            "example": "admin"
          },
          "city": {
            "type": "string"
          }
        }
      },
...
      "commonErrors": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "example": "Something went wrong"
          },
          "code": {
            "type": "number",
            "example": 400
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "Enter JWT access token",
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    }
  }
}